examples/bioevents/requirements.txt
Django==1.11.4
djangorestframework==3.6.3
django-rest-swagger==2.1.2
psycopg2==2.7.3
python-dotenv==0.6.4
